# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CND Rest API",
        "version": "7.0"
    },
    "servers": [
        {
            "url": "http://ins-2023.aml.fr/api/v7",
            "description": "REST APIs CND"
        },
        {
            "url": "https://ins-demo.cnd.info/api/v7",
            "description": "Site Demo"
        },
        {
            "url": "https://ins-bs.obamalou.fr/api/v7",
            "description": "Site Demo"
        }
    ],
    "paths": {
...
        "/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Returns a student identified by its Id",
                "operationId": "studentGetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "student id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "opt",
                        "in": "query",
                        "description": "option",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "startDate"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Students"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "operationId": "studentDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "student id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Returns a list of students",
                "operationId": "studentGet",
                "parameters": [
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "date calcul age CND",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "opt",
                        "in": "query",
                        "description": "option",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "birthdayValidated",
                                "photoValidated",
                                "member",
                                "boardMember",
                                "usedOnly",
                                "followUp",
                                "allFollowUp",
                                "boardMemberFollowUp",
                                "duplicates",
                                "csv"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Students"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/students/{id}/{type}/{value}": {
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Updates a student identified by its Id, with the type and its new value",
                "operationId": "studentPut",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "student id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "operation type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "valCni",
                                "valPhoto",
                                "valCard",
                                "valMinor",
                                "valMajor",
                                "cndPaid",
                                "indPaid",
                                "duoPaid",
                                "grpPaid",
                                "comment"
                            ]
                        }
                    },
                    {
                        "name": "value",
                        "in": "path",
                        "description": "operation value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "properties": {
                    "Error": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Common": {
                "title": "Common",
                "properties": {
                    "Common": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "method": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "value": {
                                    "type": "string"
                                },
                                "count": {
                                    "type": "integer"
                                },
                                "type": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Header": {
                "title": "Header",
                "properties": {
                    "objectType": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "integer"
                    },
                    "statusMessage": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Response": {
                "title": "Response",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Header"
                    },
                    {
                        "$ref": "#/components/schemas/Error"
                    },
                    {
                        "$ref": "#/components/schemas/Common"
                    }
                ]
            },
...
        }
    },
...
}